refactor: modularize modal component & reduce global state variables
