+++ src/constants/urls.ts
@@ -22,6 +22,8 @@ export const GOOGLE_HOMEPAGE = "https://www.google.com";
 
 export const CHROME_SHORTCUTS = "chrome://extensions/shortcuts";
 
+export const CHROME_NEW_TAB = "chrome://newtab";
+
 // LICENSE
 const GITHUB_BASE_URL = "https://github.com";
