+++ src/pages/Settings/Backup.tsx
@@ -1,0 +1,8 @@
+export default function Backup(): JSX.Element {
+  return (
+    <div>
+      <p>Automatic Export every X minutes to folder, keeping most recent Y backups</p>
+      <p>Automatic Sync every X minutes</p>
+    </div>
+  );
+}
