+++ src/pages/Settings/Filter.tsx
@@ -1,0 +1,10 @@
+export default function Filter(): JSX.Element {
+  return (
+    <div>
+      <p>
+        The following websites are excluded from TabMerger (default chrome:// and new tab), RegExp to filter for
+        https?://www.
+      </p>
+    </div>
+  );
+}
