+++ src/pages/Settings/Theme.tsx
@@ -1,0 +1,7 @@
+export default function Theme(): JSX.Element {
+  return (
+    <div>
+      <p>Dark/Light Mode with example</p>
+    </div>
+  );
+}
